backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic[email]==2.5.0
PyJWT==2.8.0
python-dotenv==1.0.0
pydantic-settings==2.1.0

